Separate file for data structure refactor
